Automate Transfers: Only email on USER_INPUT once
refs #7374

Track the current microservice a unit is waiting on. Use that to pass a
parameter to the user-input scripts of whether or not this is the first
time the script has been called for this unit on this wait point.  Update
send_email to only email if it is the first time.