transfer: check script extension against configured list before exec
The run_scripts function in transfer.py was very trusting about what should
be executable. Previously, if the execute bit was set on a file, it was
considered executable, which works fine unless someone/something has set
the permissions wrong on the file(s). In particular, we want to avoid
attempting to execute the processing XML files, which we may reasonably
expect to be alongside the executable scripts in the pre-transfer folder.

This commit adds a "whitelist" of extensions that a script should have.
The default is ".py:.sh" (colon-separated) but this can be changed using
the scriptextensions option in the config file. If the whitelist is defined
then files without these extensions will not be considered for execution.

I've also made a change to start_transfer: we now recover from any error
when executing run_scripts, so that the pid lock file is not left lying
around (which happened previously when the script crashed out).