Creates a bulk reingest mechanism
This is a script for the Canadian Center for Architects. The purpose
is to bulk reingest AIPs in Archivematica from a single pipeline.
The script can be run using the same mechanism as others in the
automation tools repository, i.e. running a crontab task every (x)
minutes. In support of this script, PR #57 was submitted to enhance
the amclient.py script's capabilities.

* Resolves #81 create a bulk-reingest mechanism
* Resolves #72 write-mode when opening the PID file