Provide option to delete transfer src post-ingest
This commit adds a `--delete-on-complete` flag so that when
`transfer.py` has access to the transfer source the user can also
request that it be deleted so that files are not unecessarily kept
around. The default for the script is 'not' to delete the source
location so that users have to explicitly request the removal of
these files.