Correct the behavior of the transfer path
The previous commit broke how the transfer path was utilised by the
script, namely, a non-absolute path was sent to the pre-transfers
scripts which saw them manipulate the automation tools folder instead
of the transfer folder. The previous functionality has been
re-implemented and tests created to demonstrate the return of a
tuple containing the transfer name and the transfer path.
Additionally, the automation-tools will verify that the transfer_path
can be discovered, because if it doesn't exist for any reasson, e.g.
the scripts are being run remotely, or have been configured
incorrectly, then we do not want to run the scripts at all.