Reenable transfer.py's use of target name
This transfer makes sure that the correct information is written to
the transfer database to ensure that automation tools works as it
did previously. Automation tools need to be able to locate a transfer
in its database by name, and not path, which the previous commits
introduced.

Further integration unit tests have been introduced to preserve the
integrity of the workflow through further changes to this utility.