Add additional functions to the amclient module
In this commit we add three module functions as well as tests and
CLI commands to amclient.py. AMClient interfaces with the
Archivematica API. The added endpoints are:

* Retrieve storage service locations endpoint
* V2 create package endpoint
* Retrieve file from package endpoint
* We also add a missing download-aip endpoint

A minor change has been made to pre-processing CLI parameters where
path strings were encoded early, but this led to a complicated dance
of decoding and encoding within module functions to make the code
compatible with either Python 2 or 3 libraries. For now the consuming
functions will be responsible for the encoding and decoding of paths,
receiving them as strings, before talking to the storage service or
dashboard as required.