Add DIP upload to SS script
Uploads a local DIP to a DIP storage location in an SS instance.
Requires access to a pipeline's currently processing location path (the
shared path), to move the DIP folder in there and send a requests to the
Storage Service to process that DIP and create a relationship with the
AIP from where it was created.

Add METS type argument to `create_dip` script to choose between a
generated METS file that matches AtoM requirements for DIP upload and
copying the AIP's METS file directly, like the Storage Service does, to
be able to move the DIP to the SS. Defaults to AtoM METS for backwards
compatibility.

Allow to upload DIPs after creation in `create_dips_job`. Accept
different subsets of parameters and optionally upload the created DIPs
to AtoM or the Storage Service. Extend this script to make use of its
tracking functionality to create DIPs from all AIPs in a SS location.
Optionally delete the created DIPs after upload.

Use AIP folder name for DIP folder on creation to match what's currently
being done for DIPs stored in the SS.